flaskandmnist/requirements.txt
absl-py==0.7.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
beautifulsoup4==4.6.0
blinker==1.4
boto3==1.9.50
botocore==1.12.52
Brlapi==0.6.6
canvas==0.2
cassandra-cql==0.9.0
cassandra-driver==3.18.0
cassandra-flask-sessions==0.6.0
cassandras3==0.1.6
certifi==2018.1.18
chardet==3.0.4
click==6.2
command-not-found==0.3
cql==1.4.0
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
DateTime==4.3
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
docutils==0.12
eight==0.3.3
feedparser==5.2.1
Flask==1.1.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
funcsigs==1.0.2
future==0.15.2
futures==3.0.5
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
itsdangerous==1.1.0
javascript==2015.32.59
Jinja2==2.10.1
jmespath==0.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
mock==2.0.0
netifaces==0.10.4
numpy==1.17.0
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
opencv-python==4.1.0.25
pandas==0.25.0
pbr==1.9.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.9.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
s3transfer==0.1.13
scipy==1.3.1
SecretStorage==2.3.1
sh==1.12.13
simplejson==3.13.2
six==1.11.0
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
thrift==0.11.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
webencodings==0.5
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
xkit==0.0.0
zope.interface==4.3.2
